+++ misc/services/mobile-webui/mobile-webui-frontend/src/apps/pos/containers/order_panel/CurrentOrder.jsx
@@ -1,12 +1,10 @@
-import React, { useEffect, useRef } from 'react';
-import cx from 'classnames';
+import React from 'react';
 import { setSelectedOrderLineAction, useCurrentOrderOrNew, usePOSTerminal } from '../../actions';
 import CurrentOrderActions from './CurrentOrderActions';
-import PropTypes from 'prop-types';
-import { formatQtyToHumanReadableStr } from '../../../../utils/qtys';
 import { formatAmountToHumanReadableStr } from '../../../../utils/money';
 import { useDispatch } from 'react-redux';
 import './CurrentOrder.scss';
+import { OrderLine } from './OrderLine';
 
 const CurrentOrder = () => {
   const dispatch = useDispatch();
@@ -64,76 +62,5 @@ const CurrentOrder = () => {
     </div>
   );
 };
-
-//
-//
-//
-//
-//
-
-const OrderLine = ({
-  /*uuid,*/ productName,
-  qty,
-  uom,
-  catchWeight,
-  catchWeightUom,
-  currencySymbol,
-  price,
-  amount,
-  pricePrecision,
-  currencyPrecision,
-  selected,
-  onClick,
-}) => {
-  const elementRef = useRef();
-  const amountStr =
-    amount != null
-      ? formatAmountToHumanReadableStr({ amount: amount, currency: currencySymbol, precision: currencyPrecision })
-      : '';
-  const qtyStr = formatQtyToHumanReadableStr({ qty, uom });
-  const catchWeightStr =
-    catchWeightUom != null ? formatQtyToHumanReadableStr({ qty: catchWeight, uom: catchWeightUom }) : null;
-  const priceUom = catchWeightUom ? catchWeightUom : uom;
-  const priceStr =
-    formatAmountToHumanReadableStr({ amount: price, currency: currencySymbol, precision: pricePrecision }) +
-    '/' +
-    priceUom;
-  const description = `${qtyStr}${catchWeightStr ? ', ' + catchWeightStr : ''} at ${priceStr}`;
-
-  useEffect(() => {
-    if (selected && elementRef?.current?.scrollIntoView) {
-      elementRef.current.scrollIntoView({ behaviour: 'smooth', block: 'end', inline: 'end' });
-    }
-  }, [selected]);
-
-  return (
-    <div className={cx('line', { 'line-selected': selected })} ref={elementRef} onClick={onClick}>
-      <div className="main">
-        <div className="productName">{productName}</div>
-        <div className="amount">{amountStr}</div>
-      </div>
-      <div className="description">{description}</div>
-    </div>
-  );
-};
-
-OrderLine.propTypes = {
-  productName: PropTypes.string.isRequired,
-  qty: PropTypes.number.isRequired,
-  uom: PropTypes.string.isRequired,
-  currencySymbol: PropTypes.string,
-  price: PropTypes.number,
-  amount: PropTypes.number,
-  pricePrecision: PropTypes.number,
-  currencyPrecision: PropTypes.number,
-  selected: PropTypes.bool.isRequired,
-  onClick: PropTypes.func.isRequired,
-};
-
-//
-//
-//
-//
-//
 
 export default CurrentOrder;
